[cheng.gao] #calltrace: adjust folder structure
